requirements.txt
numpy==1.21.6
pandas==1.3.5
PyYAML==6.0
recbole==1.0.1
torch==1.13.0